requirements.txt
python-telegram-bot==20.7
motor==3.3.2
pymongo==4.6.3
python-dotenv==1.0.1
anthropic==0.25.1
resend==0.7.2
apscheduler==3.10.4
aiohttp==3.9.3
beautifulsoup4==4.12.3
thefuzz==0.22.1
python-Levenshtein==0.25.1
httpx
nest-asyncio
